10er

+++ ft_printf.h
@@ -1,16 +1,11 @@
 #ifndef FT_PRINTF_H
 # define FT_PRINTF_H
 
-# include "../ft_libft/libft.h"
 # include <stdarg.h>
-
-
-
-
 
 int	ft_putchar(char c, int fd);
 int	ft_printf(char *str, ...);
-size_t    ft_putstr(char *s, int fd);
+int    ft_putstr(char *s, int fd);
 int     ft_print_putnbr(long long n);
 int ft_putnbr_hex_lower(unsigned long n);
 int ft_putnbr_hex_upper(unsigned int n);
