+++ ft_printstr.c
@@ -1,8 +1,8 @@
 #include "ft_printf.h"
 
-size_t    ft_putstr(char *s, int fd)
+int    ft_putstr(char *s, int fd)
 {
-        size_t  i;
+        int  i;
 
         i = 0;
         while (i < ft_strlen(s))
